@@ -2,7 +2,15 @@ cmake_minimum_required(VERSION 3.0)
 
 project(endless_mobile)
 
-set(CMAKE_CXX_STANDARD 17 CACHE STRING "")
+# We aren't using a new enough cmake to use c++20 this way. Hardwire the
+# standard directly instead.
+# set(CMAKE_CXX_STANDARD 20 CACHE STRING "")
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++2a")
+set(CMAKE_C_FLAGS "${CMAKE_CXX_FLAGS} -std=c11")
+
+# SDL requires c89 compliance, but their android code is written against
+# C99. This isn't fixed until 2.30
+set(HAVE_GCC_WDECLARATION_AFTER_STATEMENT 0)
 
 option(ENABLE_SHARED "" FALSE)
 option(ENABLED_STATIC "" TRUE)
