@@ -223,7 +223,7 @@ bool Dialog::AllowsFastForward() const noexcept
 bool Dialog::KeyDown(SDL_Keycode key, Uint16 mod, const Command &command, bool isNewPress)
 {
 	auto it = KEY_MAP.find(key);
-	bool isCloseRequest = key == SDLK_ESCAPE || key == SDLK_AC_BACK || (key == 'w' && (mod & (KMOD_CTRL | KMOD_GUI)));
+	bool isCloseRequest = key == SDLK_ESCAPE || (key == 'w' && (mod & (KMOD_CTRL | KMOD_GUI)));
 	if((it != KEY_MAP.end() || (key >= ' ' && key <= '~')) && !isMission && (intFun || stringFun) && !isCloseRequest)
 	{
 		int ascii = (it != KEY_MAP.end()) ? it->second : key;
@@ -275,21 +275,11 @@ bool Dialog::KeyDown(SDL_Keycode key, Uint16 mod, const Command &command, bool i
 			if(!isOkDisabled)
 			{
 				DoCallback();
-				if (intFun || stringFun)
-				{
-					SDL_StopTextInput();
-				}
 				GetUI()->Pop(this);
 			}
 		}
 		else
-		{
-			if (intFun || stringFun)
-			{
-				SDL_StopTextInput();
-			}
 			GetUI()->Pop(this);
-		}
 	}
 	else if((key == 'm' || command.Has(Command::MAP)) && system && player)
 		GetUI()->Push(new MapDetailPanel(*player, system));
@@ -320,12 +310,6 @@ bool Dialog::Click(int x, int y, int clicks)
 			okIsActive = false;
 			return DoKey(SDLK_RETURN);
 		}
-	}
-
-g	if (intFun || stringFun)
-	{
-		// Clicked on edit field. popup touchscreen keyboard if needed.
-		SDL_StartTextInput();
 	}
 
 	return true;
