@@ -34,6 +34,8 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 
 using namespace std;
 
+int Panel::zoneFingerId = -1;
+int Panel::panelFingerId = -1;
 
 
 Panel::Panel() noexcept
@@ -79,6 +81,10 @@ void Panel::Step()
 // drawing any of the panels under it.
 bool Panel::IsFullScreen() const noexcept
 {
+	for(auto it = children.rbegin(); it != children.rend(); ++it)
+		if ((*it)->IsFullScreen())
+			return true;
+
 	return isFullScreen;
 }
 
@@ -88,6 +94,9 @@ bool Panel::IsFullScreen() const noexcept
 // passed to any panel under it. By default, all panels do this.
 bool Panel::TrapAllEvents() const noexcept
 {
+	for(auto it = children.rbegin(); it != children.rend(); ++it)
+		if ((*it)->IsFullScreen())
+			return true;
 	return trapAllEvents;
 }
 
@@ -202,12 +211,6 @@ bool Panel::ZoneMouseDown(const Point &point, int id)
 // so, apply that zone's action and return true.
 bool Panel::ZoneFingerDown(const Point &point, int id)
 {
-	for(auto it = children.rbegin(); it != children.rend(); ++it)
-	{
-		if((*it)->ZoneClick(point))
-			return true;
-	}
-
 	for(const Zone &zone : zones)
 	{
 		if(zone.Contains(point))
@@ -368,6 +371,134 @@ bool Panel::DoRelease(int x, int y)
 bool Panel::DoScroll(double dx, double dy)
 {
 	return EventVisit(&Panel::Scroll, dx, dy);
+}
+
+
+
+bool Panel::DoFingerDown(int x, int y, int fid, int clicks)
+{
+	// Order:
+	//   0. Children first.
+	//   1. Zones (these will be buttons)
+	//   2. Finger down events (this will be game controls)
+	//      2.5 Trigger a hover as well, as some ui's use this to
+	//          determine where a drag begins from.
+	//   3. Clicks (fallback to mouse click)
+	for(auto it = children.rbegin(); it != children.rend(); ++it)
+		if((*it)->DoFingerDown(x, y, fid, clicks))
+			return true;
+
+	if(ZoneFingerDown(Point(x, y), fid))
+	{
+		zoneFingerId = fid;
+		return true;
+	}
+	if(FingerDown(x, y, fid))
+		return true;
+	Hover(x, y);
+	if(DoClick(x, y, clicks))
+	{
+		panelFingerId = fid;
+		return true;
+	}
+	return false;
+}
+
+
+
+bool Panel::DoFingerMove(int x, int y, double dx, double dy, int fid)
+{
+	// Order:
+	//   0. Children
+	//   1. FingerMove events (These will be game controls)
+	//   2. Drag (ui events)
+	for(auto it = children.rbegin(); it != children.rend(); ++it)
+		if((*it)->DoFingerMove(x, y, dx, dy, fid))
+			return true;
+
+	if(FingerMove(x, y, fid))
+		return true;
+	if(fid == panelFingerId)
+		return Drag(dx, dy);
+	return false;
+}
+
+
+
+bool Panel::DoFingerUp(int x, int y, int fid)
+{
+	// Order:
+	//   0. Children
+	//   1. Zones (these will be buttons)
+	//   2. Finger down events (this will be game controls)
+	//   3. Clicks (fallback to mouse click)
+	for(auto it = children.rbegin(); it != children.rend(); ++it)
+		if((*it)->DoFingerUp(x, y, fid))
+			return true;
+
+	if(fid == zoneFingerId)
+	{
+		zoneFingerId = -1;
+		if (HasZone(Point(x, y)))
+			return true;
+	}
+	if(FingerUp(x, y, fid))
+		return true;
+	if (fid == panelFingerId)
+	{
+		panelFingerId = -1;
+		return Release(x, y);
+	}
+	return false;
+}
+
+
+
+bool Panel::DoGesture(Gesture::GestureEnum gesture)
+{
+	return EventVisit(&Panel::Gesture, gesture);
+}
+
+
+
+bool Panel::DoControllersChanged()
+{
+	return EventVisit(&Panel::ControllersChanged);
+}
+
+
+
+bool Panel::DoControllerButtonDown(SDL_GameControllerButton button)
+{
+	return EventVisit(&Panel::ControllerButtonDown, button);
+}
+
+
+
+bool Panel::DoControllerButtonUp(SDL_GameControllerButton button)
+{
+	return EventVisit(&Panel::ControllerButtonUp, button);
+}
+
+
+
+bool Panel::DoControllerAxis(SDL_GameControllerAxis axis, int position)
+{
+	return EventVisit(&Panel::ControllerAxis, axis, position);
+}
+
+
+
+bool Panel::DoControllerTriggerPressed(SDL_GameControllerAxis axis, bool positive)
+{
+	return EventVisit(&Panel::ControllerTriggerPressed, axis, positive);
+}
+
+
+
+bool Panel::DoControllerTriggerReleased(SDL_GameControllerAxis axis, bool positive)
+{
+	return EventVisit(&Panel::ControllerTriggerReleased, axis, positive);
 }
 
 
