@@ -240,7 +240,7 @@ void ShipyardPanel::Buy(bool onlyOwned)
 	if(licenseCost < 0)
 		return;
 
-	modifier = stoi(selected_quantity.GetSelected());
+	modifier = stoi(selected_quantity->GetSelected());
 	string message;
 	if(licenseCost)
 		message = "Note: you will need to pay " + Format::CreditString(licenseCost)
